requirements.txt
annotated-types==0.6.0
pydantic-core==2.16.3
pydantic==2.6.4
python-dotenv==1.0.1
typing-extensions==4.10.0
xmltodict==0.13.0